Feature/board

+++ src/main/java/fc/side/fastboard/board/repository/BoardRepository.java
@@ -7,7 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface BoardRepository extends JpaRepository<Board, Integer> {
-
+public interface BoardRepository extends JpaRepository<Board, Long> {
 
 }
